update value types for monetary properties

+++ BusinessObjects/Entities/AuctionFashionItem.cs
@@ -3,7 +3,7 @@
 public class AuctionFashionItem : FashionItem
 {
     public int Duration { get; set; }
-    public decimal InitialPrice { get; set; }
+    public int InitialPrice { get; set; }
     public string AuctionItemStatus { get; set; }
 
     public ICollection<Auction> Auctions = new List<Auction>();
