+++ BusinessObjects/Entities/FashionItem.cs
@@ -7,11 +7,11 @@ public  class FashionItem
     [Key]
     public Guid ItemId { get; set; }
     public string Type { get; set; }
-    public decimal SellingPrice { get; set; }
+    public int SellingPrice { get; set; }
     public string Name { get; set; }
     public string Note { get; set; }
     public int Quantity { get; set; }
-    public decimal? Value { get; set; }
+    public int? Value { get; set; }
     
     public string Condition { get; set; }
     public Request Request { get; set; }
